implement correct scalar type
