+++ src/utils/protocol-consts.js
@@ -1,0 +1,12 @@
+// Docs: https://github.com/codaco/Network-Canvas/wiki/Variable-Types
+export const VariableType = Object.freeze({
+  boolean: 'boolean',
+  text: 'text',
+  number: 'number',
+  datetime: 'datetime',
+  ordinal: 'ordinal',
+  categorical: 'categorical',
+  layout: 'layout',
+  scalar: 'scalar',
+  datetime: 'datetime',
+});
